tasks/ppc-drovosec/server/requirements.txt
aiortc
fastapi
pillow
numpy
pydantic
redis
